% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorist-package.R
\docType{package}
\name{errorist-package}
\alias{errorist}
\alias{errorist-package}
\title{errorist: Automatically Search Errors or Warnings}
\description{
This package provides shims to obtain errors and warnings that
occur during execution of code to automatically search the contents.
}
\details{
By default, \code{errorist} searches Google for both warnings and errors.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
