% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorist-package.R
\docType{package}
\name{errorist-package}
\alias{errorist}
\alias{errorist-package}
\title{errorist: Automatically Search Errors or Warnings}
\description{
Provides environment hooks that obtain errors and warnings which
    occur during the execution of code to automatically search for solutions.
}
\details{
By default, \code{errorist} searches Google for both warnings and errors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-assist/errorist}
  \item Report bugs at \url{https://github.com/r-assist/errorist/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

}
